+++ CHANGES.md
@@ -1,10 +1,9 @@
 ## dev
 
 * WIP -- Update to [ProseMirror 0.8.0](https://discuss.prosemirror.net/t/release-0-8-0/336)
-* BREAKING -- changed build and dist directory
+* BREAKING -- changed dist directory
   * `dist/build.js` for all-inclusive built library that exposes `window.TheGridEd`
   * `dist/ed.js` (package main) for the React component
-* BREAKING -- `mountApp` does not return the mounted Ed instance. Now you have to pass a `ref` callback function with the props. This is synchronous for now, but [might be async](https://facebook.github.io/react/docs/top-level-api.html#reactdom.render) in the future.
 
 ## 0.16.0 - 2016-06-14